requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
asyncpg==0.27.0
injectable==3.4.3
pydantic==1.10.7
python-dotenv==0.21.1
sqlalchemy==1.4.46
gspread==5.7.0
google-auth==2.16.0
alembic==1.9.4
